Select, pretty textarea working.
