+++ demo/index.html
@@ -28,7 +28,9 @@
     var fields = [
       {label: 'Name', type: 'unicode', key: 'name'},
       {label: 'Note', type: 'text', key: 'note'},
-      {id: 'silly', type: 'copy', help_text_html: 'What the hey?'}
+      {id: 'silly', type: 'copy', help_text_html: 'What the hey?'},
+      {label: 'Mood', type: 'select', key: 'mood', choices: ['happy', 'sad']},
+      {label: 'Groceries', type: 'pretty-textarea', key: 'groceries', replaceChoices: ['tacos', 'nachos', 'bread', 'milk', 'eggs']}
     ];
 
     var value = {};
