+++ lib/components/fields/list.js
@@ -15,7 +15,7 @@ module.exports = function (plugin) {
 
   plugin.exports = React.createClass({
 
-    displayName: plugin.name,
+    displayName: 'List',
 
     mixins: [plugin.require('mixin.field')],
