+++ lib/components/fields/object.js
@@ -25,7 +25,7 @@ module.exports = function (plugin) {
 
   plugin.exports = React.createClass({
 
-    displayName: plugin.name,
+    displayName: 'Object',
 
     mixins: [plugin.require('mixin.field')],
