+++ lib/components/fields/pretty-textarea.js
@@ -1,0 +1,777 @@
+// # component.pretty-textarea
+
+/*
+Textarea that will display highlights behind "tags". Tags currently mean text
+that is enclosed in braces like `{{foo}}`. Tags are replaced with labels if
+available or humanized.
+
+This component is quite complicated because:
+- We are displaying text in the textarea but have to keep track of the real
+  text value in the background. We can't use a data attribute, because it's a
+  textarea, so we can't use any elements at all!
+- Because of the hidden data, we also have to do some interception of
+  copy, which is a little weird. We intercept copy and copy the real text
+  to the end of the textarea. Then we erase that text, which leaves the copied
+  data in the buffer.
+- React loses the caret position when you update the value to something
+  different than before. So we have to retain tracking information for when
+  that happens.
+- Because we monkey with copy, we also have to do our own undo/redo. Otherwise
+  the default undo will have weird states in it.
+
+So good luck!
+*/
+
+'use strict';
+
+var React = require('react/addons');
+var R = React.DOM;
+var _ = require('underscore');
+var cx = React.addons.classSet;
+
+var utils = require('../../utils');
+
+var noBreak = function (value) {
+  return value.replace(/ /g, '\u00a0');
+};
+
+var LEFT_PAD = '\u00a0\u00a0';
+// Why this works, I'm not sure.
+var RIGHT_PAD = '  '; //'\u00a0\u00a0';
+
+var idPrefixRegEx = /^[0-9]+__/;
+
+// Zapier specific stuff. Make a plugin for this later.
+var removeIdPrefix = function (key) {
+  if (idPrefixRegEx.test(key)) {
+    return key.replace(idPrefixRegEx, '');
+  }
+  return key;
+};
+
+var positionInNode = function (position, node) {
+  var rect = node.getBoundingClientRect();
+  if (position.x >= rect.left && position.x <= rect.right) {
+    if (position.y >= rect.top && position.y <= rect.bottom) {
+      return true;
+    }
+  }
+};
+
+// Wrap a text value so it has a type. For parsing text with tags.
+var textPart = function (value, type) {
+  type = type || 'text';
+  return {
+    type: type,
+    value: value
+  };
+};
+
+// Parse text that has tags like {{tag}} into text and tags.
+var parseTextWithTags = function (value) {
+  value = value || '';
+  var parts = value.split(/{{(?!{)/);
+  var frontPart = [];
+  if (parts[0] !== '') {
+    frontPart = [
+    textPart(parts[0])
+    ];
+  }
+  parts = frontPart.concat(
+    parts.slice(1).map(function (part) {
+      if (part.indexOf('}}') >= 0) {
+        return [
+        textPart(part.substring(0, part.indexOf('}}')), 'tag'),
+        textPart(part.substring(part.indexOf('}}') + 2))
+        ];
+      } else {
+        return textPart('{{' + part, 'text');
+      }
+    })
+  );
+  return [].concat.apply([], parts);
+};
+
+
+module.exports = React.createClass({
+
+  displayName: 'PrettyTextarea',
+
+  mixins: [require('../../mixins/field'), require('../../mixins/undo-stack'), require('../../mixins/resize')],
+
+  //
+  // getDefaultProps: function () {
+  //   return {
+  //     className: plugin.config.className
+  //   };
+  // },
+
+  getReplaceState: function (props) {
+    var replaceChoices = props.config.getFieldReplaceChoices(this.props.field);
+    var replaceChoicesLabels = {};
+    replaceChoices.forEach(function (choice) {
+      replaceChoicesLabels[choice.value] = choice.label;
+    });
+    return {
+      replaceChoices: replaceChoices,
+      replaceChoicesLabels: replaceChoicesLabels
+    };
+  },
+
+  getInitialState: function () {
+    var replaceState = this.getReplaceState(this.props);
+
+    return {
+      undoDepth: 100,
+      isChoicesOpen: false,
+      hoverPillRef: null,
+      replaceChoices: replaceState.replaceChoices,
+      replaceChoicesLabels: replaceState.replaceChoicesLabels
+    };
+  },
+
+  componentWillReceiveProps: function (newProps) {
+    this.setState(this.getReplaceState(newProps));
+  },
+
+  componentWillMount: function () {
+    // Not quite state, this is for tracking selection info.
+    this.tracking = {};
+
+    var parts = parseTextWithTags(this.props.value);
+    var tokens = this.tokens(parts);
+    var indexMap = this.indexMap(tokens);
+
+    this.tracking.pos = indexMap.length;
+    this.tracking.range = 0;
+    this.tracking.tokens = tokens;
+    this.tracking.indexMap = indexMap;
+  },
+
+  getStateSnapshot: function () {
+    return {
+      value: this.props.value,
+      pos: this.tracking.pos,
+      range: this.tracking.range
+    };
+  },
+
+  setStateSnapshot: function (snapshot) {
+    this.tracking.pos = snapshot.pos;
+    this.tracking.range = snapshot.range;
+    this.onChangeValue(snapshot.value);
+  },
+
+  // Turn into individual characters and tags
+  tokens: function (parts) {
+    return [].concat.apply([], parts.map(function (part) {
+      if (part.type === 'tag') {
+        return part;
+      } else {
+        return part.value.split('');
+      }
+    }));
+  },
+
+  // Map each textarea index back to a token
+  indexMap: function (tokens) {
+    var indexMap = [];
+    _.each(tokens, function (token, tokenIndex) {
+      if (token.type === 'tag') {
+        var label = LEFT_PAD + noBreak(this.prettyLabel(token.value)) + RIGHT_PAD;
+        var labelChars = label.split('');
+        _.each(labelChars, function () {
+          indexMap.push(tokenIndex);
+        });
+      } else {
+        indexMap.push(tokenIndex);
+      }
+    }.bind(this));
+    return indexMap;
+  },
+
+  // Make highlight scroll match textarea scroll
+  onScroll: function () {
+    this.refs.highlight.getDOMNode().scrollTop = this.refs.content.getDOMNode().scrollTop;
+    this.refs.highlight.getDOMNode().scrollLeft = this.refs.content.getDOMNode().scrollLeft;
+  },
+
+  // Given some postion, return the token index (position could be in the middle of a token)
+  tokenIndex: function (pos, tokens, indexMap) {
+    if (pos < 0) {
+      pos = 0;
+    } else if (pos >= indexMap.length) {
+      return tokens.length;
+    }
+    return indexMap[pos];
+  },
+
+  onChange: function (event) {
+    //console.log('change:', event.target.value);
+
+    var node = event.target;
+
+    // Tracking is holding previous position and range
+    var prevPos = this.tracking.pos;
+    var prevRange = this.tracking.range;
+
+    // New position
+    var pos = node.selectionStart;
+
+    // Going to mutate the tokens.
+    var tokens = this.tracking.tokens;
+
+    // Using the previous position and range, get the previous token position
+    // and range
+    var prevTokenIndex = this.tokenIndex(prevPos, tokens, this.tracking.indexMap);
+    var prevTokenEndIndex = this.tokenIndex(prevPos + prevRange, tokens, this.tracking.indexMap);
+    var prevTokenRange = prevTokenEndIndex - prevTokenIndex;
+
+    // Wipe out any tokens in the selected range because the change would have
+    // erased that selection.
+    if (prevTokenRange > 0) {
+      tokens.splice(prevTokenIndex, prevTokenRange);
+      this.tracking.indexMap = this.indexMap(tokens);
+    }
+
+    // If cursor has moved forward, then text was added.
+    if (pos > prevPos) {
+      var addedText = node.value.substring(prevPos, pos);
+      // Insert the text into the tokens.
+      tokens.splice(prevTokenIndex, 0, addedText);
+    // If cursor has moved backward, then we deleted (backspaced) text
+    } if (pos < prevPos) {
+      var token = this.tokenAt(pos);
+      var tokenBefore = this.tokenBefore(pos);
+      // If we moved back onto a token, then we should move back to beginning
+      // of token.
+      if (token === tokenBefore) {
+        pos = this.moveOffTag(pos, tokens, this.indexMap(tokens), -1);
+      }
+      var tokenIndex = this.tokenIndex(pos, tokens, this.tracking.indexMap);
+      // Now we can remove the tokens that were deleted.
+      tokens.splice(tokenIndex, prevTokenIndex - tokenIndex);
+    }
+
+    // Convert tokens back into raw value with tags. Newly formed tags will
+    // become part of the raw value.
+    var rawValue = this.rawValue(tokens);
+
+    this.tracking.pos = pos;
+    this.tracking.range = 0;
+
+    // Set the value to the new raw value.
+    this.onChangeValue(rawValue);
+
+    this.snapshot();
+  },
+
+  componentDidUpdate: function () {
+    var value = this.props.value || '';
+    var parts = parseTextWithTags(value);
+    this.tracking.tokens = this.tokens(parts);
+    this.tracking.indexMap = this.indexMap(this.tracking.tokens);
+
+    var pos = this.normalizePosition(this.tracking.pos);
+    var range = this.tracking.range;
+    var endPos = this.normalizePosition(pos + range);
+    range = endPos - pos;
+
+    this.tracking.pos = pos;
+    this.tracking.range = range;
+
+    if (document.activeElement === this.refs.content.getDOMNode()) {
+      // React can lose the selection, so put it back.
+      this.refs.content.getDOMNode().setSelectionRange(pos, pos + range);
+    }
+  },
+
+  // Get the label for a key.
+  prettyLabel: function (key) {
+    if (this.state.replaceChoicesLabels[key]) {
+      return this.state.replaceChoicesLabels[key];
+    }
+    var cleaned = removeIdPrefix(key);
+    return this.props.config.humanize(cleaned);
+  },
+
+  // Given the actual value of the field (with tags), get the plain text that
+  // should show in the textarea.
+  plainValue: function (value) {
+    var parts = parseTextWithTags(value);
+    return parts.map(function (part) {
+      if (part.type === 'text') {
+        return part.value;
+      } else {
+        return LEFT_PAD + noBreak(this.prettyLabel(part.value)) + RIGHT_PAD;
+      }
+    }.bind(this)).join('');
+  },
+
+  // Given the actual value of the field (with tags), get the html used to
+  // highlight the labels.
+  prettyValue: function (value) {
+    var parts = parseTextWithTags(value);
+    return parts.map(function (part, i) {
+      if (part.type === 'text') {
+        if (i === (parts.length - 1)) {
+          if (part.value[part.value.length - 1] === '\n') {
+            return part.value + '\u00a0';
+          }
+        }
+        return part.value;
+      } else {
+        // Make a pill
+        var pillRef = 'prettyPart' + i;
+        var className = 'pretty-part';
+        if (this.state.hoverPillRef && pillRef === this.state.hoverPillRef) {
+          className += ' pretty-part-hover';
+        }
+        return R.span({key: i, className: className, ref: pillRef, 'data-pretty': true, 'data-ref': pillRef},
+          R.span({className: 'pretty-part-left'}, LEFT_PAD),
+          R.span({className: 'pretty-part-text'}, noBreak(this.prettyLabel(part.value))),
+          R.span({className: 'pretty-part-right'}, RIGHT_PAD)
+        );
+      }
+    }.bind(this));
+  },
+
+  // Given the tokens for a field, get the actual value of the field (with
+  // tags)
+  rawValue: function (tokens) {
+    return tokens.map(function (token) {
+      if (token.type === 'tag') {
+        return '{{' + token.value + '}}';
+      } else {
+        return token;
+      }
+    }).join('');
+  },
+
+  // Given a position, if it's on a label, get the position left or right of
+  // the label, based on direction and/or which side is closer
+  moveOffTag: function (pos, tokens, indexMap, dir) {
+    if (typeof dir === 'undefined' || dir > 0) {
+      dir = 1;
+    } else {
+      dir = -1;
+    }
+    var token;
+    if (dir > 0) {
+      token = tokens[indexMap[pos]];
+      while (pos < indexMap.length && tokens[indexMap[pos]].type === 'tag' && tokens[indexMap[pos]] === token) {
+        pos++;
+      }
+    } else {
+      token = tokens[indexMap[pos - 1]];
+      while (pos > 0 && tokens[indexMap[pos - 1]].type === 'tag' && tokens[indexMap[pos - 1]] === token) {
+        pos--;
+      }
+    }
+
+    return pos;
+  },
+
+  // Get the token at some position.
+  tokenAt: function (pos) {
+    if (pos >= this.tracking.indexMap.length) {
+      return null;
+    }
+    if (pos < 0) {
+      pos = 0;
+    }
+    return this.tracking.tokens[this.tracking.indexMap[pos]];
+  },
+
+  // Get the token immediately before some position.
+  tokenBefore: function (pos) {
+    if (pos >= this.tracking.indexMap.length) {
+      pos = this.tracking.indexMap.length;
+    }
+    if (pos <= 0) {
+      return null;
+    }
+    return this.tracking.tokens[this.tracking.indexMap[pos - 1]];
+  },
+
+  // Given a position, get a corrected position (if necessary to be
+  // corrected).
+  normalizePosition: function (pos, prevPos) {
+    if (_.isUndefined(prevPos)) {
+      prevPos = pos;
+    }
+    // At start or end, so okay.
+    if (pos <= 0 || pos >= this.tracking.indexMap.length) {
+      if (pos < 0) {
+        pos = 0;
+      }
+      if (pos > this.tracking.indexMap.length) {
+        pos = this.tracking.indexMap.length;
+      }
+      return pos;
+    }
+
+    var token = this.tokenAt(pos);
+    var tokenBefore = this.tokenBefore(pos);
+
+    // Between two tokens, so okay.
+    if (token !== tokenBefore) {
+      return pos;
+    }
+
+    var prevToken = this.tokenAt(prevPos);
+    var prevTokenBefore = this.tokenBefore(prevPos);
+
+    var rightPos = this.moveOffTag(pos, this.tracking.tokens, this.tracking.indexMap);
+    var leftPos = this.moveOffTag(pos, this.tracking.tokens, this.tracking.indexMap, -1);
+
+    if (prevToken !== prevTokenBefore) {
+      // Moved from left edge.
+      if (prevToken === token) {
+        return rightPos;
+      }
+      // Moved from right edge.
+      if (prevTokenBefore === token) {
+        return leftPos;
+      }
+    }
+
+    var newPos = rightPos;
+
+    if (pos === prevPos || pos < prevPos) {
+      if (rightPos - pos > pos - leftPos) {
+        newPos = leftPos;
+      }
+    }
+    return newPos;
+  },
+
+
+
+  onSelect: function (event) {
+    var node = event.target;
+
+    var pos = node.selectionStart;
+    var endPos = node.selectionEnd;
+
+    if (pos === endPos && this.state.hoverPillRef) {
+      var tokenAt = this.tokenAt(pos);
+      var tokenBefore = this.tokenBefore(pos);
+
+      if (tokenAt && tokenAt === tokenBefore && tokenAt.type && tokenAt.type === 'tag') {
+        // Clicked a tag.
+        var rightPos = this.moveOffTag(pos, this.tracking.tokens, this.tracking.indexMap);
+        var leftPos = this.moveOffTag(pos, this.tracking.tokens, this.tracking.indexMap, -1);
+        this.tracking.pos = leftPos;
+        this.tracking.range = rightPos - leftPos;
+        node.selectionStart = leftPos;
+        node.selectionEnd = rightPos;
+
+        this.setState({isChoicesOpen: true});
+
+        return;
+      }
+    }
+
+    pos = this.normalizePosition(pos, this.tracking.pos);
+    endPos = this.normalizePosition(endPos, this.tracking.pos + this.tracking.range);
+
+    this.tracking.pos = pos;
+    this.tracking.range = endPos - pos;
+
+    node.selectionStart = pos;
+    node.selectionEnd = endPos;
+  },
+
+  onCopy: function () {
+    var node = this.refs.content.getDOMNode();
+    var start = node.selectionStart;
+    var end = node.selectionEnd;
+    var text = node.value.substring(start, end);
+    var realStartIndex = this.tokenIndex(start, this.tracking.tokens, this.tracking.indexMap);
+    var realEndIndex = this.tokenIndex(end, this.tracking.tokens, this.tracking.indexMap);
+    var tokens = this.tracking.tokens.slice(realStartIndex, realEndIndex);
+    text = this.rawValue(tokens);
+    var originalValue = node.value;
+    node.value = node.value + text;
+    node.setSelectionRange(originalValue.length, originalValue.length + text.length);
+    window.setTimeout(function() {
+      node.value = originalValue;
+      node.setSelectionRange(start, end);
+    },0);
+  },
+
+  onCut: function () {
+    var node = this.refs.content.getDOMNode();
+    var start = node.selectionStart;
+    var end = node.selectionEnd;
+    var text = node.value.substring(start, end);
+    var realStartIndex = this.tokenIndex(start, this.tracking.tokens, this.tracking.indexMap);
+    var realEndIndex = this.tokenIndex(end, this.tracking.tokens, this.tracking.indexMap);
+    var tokens = this.tracking.tokens.slice(realStartIndex, realEndIndex);
+    text = this.rawValue(tokens);
+    var originalValue = node.value;
+    var cutValue = node.value.substring(0, start) + node.value.substring(end);
+    node.value = node.value + text;
+    node.setSelectionRange(originalValue.length, originalValue.length + text.length);
+    var cutTokens = this.tracking.tokens.slice(0, realStartIndex).concat(this.tracking.tokens.slice(realEndIndex));
+    window.setTimeout(function() {
+      node.value = cutValue;
+      node.setSelectionRange(start, start);
+      this.tracking.pos = start;
+      this.tracking.range = 0;
+      this.tracking.tokens = cutTokens;
+      this.tracking.indexMap = this.indexMap(this.tracking.tokens);
+
+      // Convert tokens back into raw value with tags. Newly formed tags will
+      // become part of the raw value.
+      var rawValue = this.rawValue(this.tracking.tokens);
+
+      // Set the value to the new raw value.
+      this.onChangeValue(rawValue);
+
+      this.snapshot();
+    }.bind(this),0);
+  },
+
+  onKeyDown: function (event) {
+
+    if (event.keyCode === 37) {
+      this.leftArrowDown = true;
+    } else if (event.keyCode === 39) {
+      this.rightArrowDown = true;
+    }
+
+    // Cmd-Z or Ctrl-Z
+    if (event.keyCode === 90 && (event.metaKey || event.ctrlKey) && !event.shiftKey) {
+      event.preventDefault();
+      this.undo();
+    // Cmd-Shift-Z or Ctrl-Y
+    } else if (
+      (event.keyCode === 89 && event.ctrlKey && !event.shiftKey) ||
+      (event.keyCode === 90 && event.metaKey && event.shiftKey)
+    ) {
+      this.redo();
+    }
+  },
+
+  onKeyUp: function (event) {
+    if (event.keyCode === 37) {
+      this.leftArrowDown = false;
+    } else if (event.keyCode === 39) {
+      this.rightArrowDown = false;
+    }
+  },
+
+  // Keep the highlight styles in sync with the textarea styles.
+  adjustStyles: function (isMount) {
+    var overlay = this.refs.highlight.getDOMNode();
+    var content = this.refs.content.getDOMNode();
+
+    var style = window.getComputedStyle(content);
+
+    var backgroundColor = style.backgroundColor;
+
+    utils.copyElementStyle(content, overlay);
+
+    overlay.style.position = 'absolute';
+    overlay.style.whiteSpace = 'pre-wrap';
+    overlay.style.color = 'rgba(0,0,0,0)';
+    overlay.style.webkitTextFillColor = 'rgba(0,0,0,0)';
+    overlay.style.resize = 'none';
+    overlay.style.borderColor = 'rgba(0,0,0,0)';
+
+    if (utils.browser.isMozilla) {
+
+      var paddingTop = parseFloat(style.paddingTop);
+      var paddingBottom = parseFloat(style.paddingBottom);
+
+      var borderTop = parseFloat(style.borderTopWidth);
+      var borderBottom = parseFloat(style.borderBottomWidth);
+
+      overlay.style.paddingTop = '0px';
+      overlay.style.paddingBottom = '0px';
+
+      overlay.style.height = (content.clientHeight - paddingTop - paddingBottom + borderTop + borderBottom) + 'px';
+      overlay.style.top = style.paddingTop;
+      overlay.style.boxShadow = 'none';
+    }
+
+    if (isMount) {
+      this.backgroundColor = backgroundColor;
+    }
+    overlay.style.backgroundColor = this.backgroundColor;
+    content.style.backgroundColor = 'rgba(0,0,0,0)';
+  },
+
+  // If the textarea is resized, need to re-sync the styles.
+  onResize: function () {
+    this.adjustStyles();
+  },
+
+  // If the window is resized, may need to re-sync the styles.
+  // Probably not necessary with element resize?
+  onResizeWindow: function () {
+    this.adjustStyles();
+  },
+
+  componentDidMount: function () {
+    this.adjustStyles(true);
+    this.setOnResize('content', this.onResize);
+    //this.setOnClickOutside('choices', this.onClickOutsideChoices);
+  },
+
+  onInsertFromSelect: function (event) {
+    if (event.target.selectedIndex > 0) {
+      var tag = event.target.value;
+      event.target.selectedIndex = 0;
+      var pos = this.tracking.pos;
+      var insertPos = this.normalizePosition(pos);
+      var tokens = this.tracking.tokens;
+      var tokenIndex = this.tokenIndex(insertPos, tokens, this.tracking.indexMap);
+      tokens.splice(tokenIndex, 0, {
+        type: 'tag',
+        value: tag
+      });
+      this.tracking.indexMap = this.indexMap(tokens);
+      var newValue = this.rawValue(tokens);
+      this.tracking.pos += this.prettyLabel(tag).length;
+      this.onChangeValue(newValue);
+    }
+  },
+
+  onInsert: function (value) {
+    var tag = value;
+    var pos = this.tracking.pos;
+    var endPos = this.tracking.pos + this.tracking.range;
+    var insertPos = this.normalizePosition(pos);
+    var endInsertPos = this.normalizePosition(endPos);
+    var tokens = this.tracking.tokens;
+    var tokenIndex = this.tokenIndex(insertPos, tokens, this.tracking.indexMap);
+    var tokenEndIndex = this.tokenIndex(endInsertPos, tokens, this.tracking.indexMap);
+    tokens.splice(tokenIndex, tokenEndIndex - tokenIndex, {
+      type: 'tag',
+      value: tag
+    });
+    this.tracking.indexMap = this.indexMap(tokens);
+    var newValue = this.rawValue(tokens);
+    this.tracking.pos += this.prettyLabel(tag).length;
+    this.onChangeValue(newValue);
+    this.setState({
+      isChoicesOpen: false
+    });
+  },
+
+  onToggleChoices: function () {
+    this.setState({
+      isChoicesOpen: !this.state.isChoicesOpen
+    });
+  },
+
+  onCloseChoices: function () {
+    this.setState({
+      isChoicesOpen: false
+    });
+  },
+
+  getCloseIgnoreNodes: function () {
+    return this.refs.toggle.getDOMNode();
+  },
+
+  onClickOutsideChoices: function () {
+    // // If we didn't click on the toggle button, close the choices.
+    // if (this.isNodeOutside(this.refs.toggle.getDOMNode(), event.target)) {
+    //   console.log('not a toggle click')
+    //   this.setState({
+    //     isChoicesOpen: false
+    //   });
+    // }
+  },
+
+  onMouseMove: function (event) {
+    // Placeholder to get at pill under mouse position. Inefficient, but not
+    // sure there's another way.
+
+    var position = {x: event.clientX, y: event.clientY};
+    var nodes = this.refs.highlight.getDOMNode().childNodes;
+    var matchedNode = null;
+    for (var i = 0; i < nodes.length; i++) {
+      var node = nodes[i];
+      if (nodes[i].getAttribute('data-pretty')) {
+        if (positionInNode(position, node)) {
+          matchedNode = node;
+          break;
+        }
+      }
+    }
+
+    if (matchedNode) {
+      if (this.state.hoverPillRef !== matchedNode.getAttribute('data-ref')) {
+        this.setState({
+          hoverPillRef: matchedNode.getAttribute('data-ref')
+        });
+      }
+    } else if (this.state.hoverPillRef) {
+      this.setState({
+        hoverPillRef: null
+      });
+    }
+  },
+
+  render: function () {
+    var config = this.props.config;
+    var field = this.props.field;
+
+    var replaceChoices = this.state.replaceChoices;
+
+    // var selectReplaceChoices = [{
+    //   value: '',
+    //   label: 'Insert...'
+    // }].concat(replaceChoices);
+
+    return config.createElement('field', {
+      field: field, plain: this.props.plain
+    }, R.div({style: {position: 'relative'}},
+
+      R.pre({
+        className: 'pretty-highlight',
+        ref: 'highlight'
+      },
+        this.prettyValue(this.props.value)
+      ),
+
+      R.textarea({
+        className: cx(_.extend({}, this.props.classes, {'pretty-content': true})),
+        ref: 'content',
+        rows: field.rows || this.props.rows,
+        name: field.key,
+        value: this.plainValue(this.props.value),
+        onChange: this.onChange,
+        onScroll: this.onScroll,
+        style: {
+          position: 'relative',
+          top: 0,
+          left: 0,
+          cursor: this.state.hoverPillRef ? 'pointer' : null
+        },
+        onKeyPress: this.onKeyPress,
+        onKeyDown: this.onKeyDown,
+        onKeyUp: this.onKeyUp,
+        onSelect: this.onSelect,
+        onCopy: this.onCopy,
+        onCut: this.onCut,
+        onMouseMove: this.onMouseMove,
+        onFocus: this.onFocus,
+        onBlur: this.onBlur
+      }),
+
+      R.a({ref: 'toggle', href: 'JavaScript' + ':', onClick: this.onToggleChoices}, 'Insert...'),
+
+      config.createElement('choices', {
+        ref: 'choices',
+        choices: replaceChoices, open: this.state.isChoicesOpen,
+        onSelect: this.onInsert, onClose: this.onCloseChoices, ignoreCloseNodes: this.getCloseIgnoreNodes
+      })
+    ));
+  }
+});
