+++ lib/components/fields/select.js
@@ -12,90 +12,100 @@ var React = require('react/addons');
 var R = React.DOM;
 var _ = require('underscore');
 
-module.exports = function (plugin) {
+module.exports = React.createClass({
 
-  plugin.exports = React.createClass({
+  displayName: 'Select',
 
-    displayName: plugin.name,
+  mixins: [require('../../mixins/field')],
 
-    mixins: [plugin.require('mixin.field')],
+  // getDefaultProps: function () {
+  //   return {
+  //     className: plugin.config.className
+  //   };
+  // },
 
-    getDefaultProps: function () {
-      return {
-        className: plugin.config.className
-      };
-    },
+  getInitialState: function () {
+    return {
+      choices: this.props.config.getFieldChoices(this.props.field)
+    };
+  },
 
-    onChange: function (event) {
-      var choiceValue = event.target.value;
-      var choiceType = choiceValue.substring(0, choiceValue.indexOf(':'));
-      if (choiceType === 'choice') {
-        var choiceIndex = choiceValue.substring(choiceValue.indexOf(':') + 1);
-        choiceIndex = parseInt(choiceIndex);
-        this.props.field.val(this.props.field.def.choices[choiceIndex].value);
-      }
-    },
+  componentWillReceiveProps: function (newProps) {
+    this.setState({
+      choices: newProps.config.getFieldChoices(newProps.field)
+    });
+  },
 
-    render: function () {
-
-      var field = this.props.field;
-      var choices = field.def.choices || [];
-
-      var choicesOrLoading;
-
-      if (choices.length === 1 && choices[0].value === '///loading///') {
-        choicesOrLoading = R.div({},
-          'Loading choices...'
-        );
-      } else {
-
-        var value = field.value !== undefined ? field.value : '';
-
-        choices = choices.map(function (choice, i) {
-          return {
-            choiceValue: 'choice:' + i,
-            value: choice.value,
-            label: choice.label
-          };
-        });
-
-        var valueChoice = _.find(choices, function (choice) {
-          return choice.value === value;
-        });
-
-        if (valueChoice === undefined) {
-
-          var label = value;
-          if (!_.isString(value)) {
-            label = JSON.stringify(value);
-          }
-          valueChoice = {
-            choiceValue: 'value:',
-            value: value,
-            label: label
-          };
-          choices = [valueChoice].concat(choices);
-        }
-
-        choicesOrLoading = R.select({
-          className: this.props.className,
-          onChange: this.onChange,
-          value: valueChoice.choiceValue,
-          onFocus: this.onFocus,
-          onBlur: this.onBlur
-        },
-          choices.map(function (choice, i) {
-            return R.option({
-              key: i,
-              value: choice.choiceValue
-            }, choice.label);
-          }.bind(this))
-        );
-      }
-
-      return plugin.component('field')({
-        field: field, plain: this.props.plain
-      }, choicesOrLoading);
+  onChange: function (event) {
+    var choiceValue = event.target.value;
+    var choiceType = choiceValue.substring(0, choiceValue.indexOf(':'));
+    if (choiceType === 'choice') {
+      var choiceIndex = choiceValue.substring(choiceValue.indexOf(':') + 1);
+      choiceIndex = parseInt(choiceIndex);
+      this.onChangeValue(this.state.choices[choiceIndex].value);
     }
-  });
-};
+  },
+
+  render: function () {
+
+    var config = this.props.config;
+    var field = this.props.field;
+    var choices = this.state.choices || [];
+
+    var choicesOrLoading;
+
+    if (choices.length === 1 && choices[0].value === '///loading///') {
+      choicesOrLoading = R.div({},
+        'Loading choices...'
+      );
+    } else {
+
+      var value = this.props.value !== undefined ? this.props.value : '';
+
+      choices = choices.map(function (choice, i) {
+        return {
+          choiceValue: 'choice:' + i,
+          value: choice.value,
+          label: choice.label
+        };
+      });
+
+      var valueChoice = _.find(choices, function (choice) {
+        return choice.value === value;
+      });
+
+      if (valueChoice === undefined) {
+
+        var label = value;
+        if (!_.isString(value)) {
+          label = JSON.stringify(value);
+        }
+        valueChoice = {
+          choiceValue: 'value:',
+          value: value,
+          label: label
+        };
+        choices = [valueChoice].concat(choices);
+      }
+
+      choicesOrLoading = R.select({
+        className: this.props.className,
+        onChange: this.onChange,
+        value: valueChoice.choiceValue,
+        onFocus: this.onFocus,
+        onBlur: this.onBlur
+      },
+        choices.map(function (choice, i) {
+          return R.option({
+            key: i,
+            value: choice.choiceValue
+          }, choice.label);
+        }.bind(this))
+      );
+    }
+
+    return config.createElement('field', {
+      config: config, field: field, plain: this.props.plain
+    }, choicesOrLoading);
+  }
+});
