+++ lib/components/fields/text.js
@@ -8,9 +8,9 @@ var cx = React.addons.classSet;
 
 module.exports = React.createClass({
 
-  displayName: 'StringField',
+  displayName: 'Text',
 
-  //mixins: [plugin.require('mixin.field')],
+  mixins: [require('../../mixins/field')],
 
   // getDefaultProps: function () {
   //   return {
@@ -21,10 +21,8 @@ module.exports = React.createClass({
 
   onChange: function (event) {
     //this.props.config.onChange(event.target.value, this.props.field, this.props.onChange);
-    this.props.onChange(event.target.value, {
-      field: this.props.field,
-      fields: [this.props.field]
-    });
+
+    this.onChangeValue(event.target.value);
   },
 
   render: function () {
