+++ lib/components/fields/unknown.js
@@ -5,11 +5,9 @@ var R = React.DOM;
 
 module.exports = React.createClass({
 
-  displayName: 'UnknownField',
+  displayName: 'Unknown',
 
   render: function () {
-    var config = this.props.config;
-
     return R.div({},
       R.div({}, 'Component not found for: '),
       R.pre({}, JSON.stringify(this.props.field))
