+++ lib/components/fields/util.js
@@ -1,0 +1,312 @@
+// # util
+
+// Some utility functions to be used by other plugins.
+
+'use strict';
+
+var _ = require('underscore');
+
+module.exports = function (plugin) {
+
+  var util = plugin.exports;
+
+  // Check if a value is "blank".
+  util.isBlank = function (value) {
+    return value === undefined || value === null || value === '';
+  };
+
+  // Set value at some path in object.
+  util.setIn = function (obj, path, value) {
+    if (_.isString(path)) {
+      path = [path];
+    }
+    if (path.length === 0) {
+      return value;
+    }
+    if (path.length === 1) {
+      obj[path[0]] = value;
+      return obj;
+    }
+    if (!obj[path[0]]) {
+      obj[path[0]] = {};
+    }
+    util.setIn(obj[path[0]], path.slice(1), value);
+    return obj;
+  };
+
+  // Remove value at path in some object.
+  util.removeIn = function (obj, path) {
+    if (_.isString(path)) {
+      path = [path];
+    }
+    if (path.length === 0) {
+      return null;
+    }
+    if (path.length === 1) {
+      if (_.isArray(obj)) {
+        if (_.isNumber(path[0])) {
+          obj.splice(path[0], 1);
+        }
+      } else if (_.isObject(obj)) {
+        delete obj[path[0]];
+      }
+      return obj;
+    }
+    if (obj[path[0]]) {
+      util.removeIn(obj[path[0]], path.slice(1));
+    }
+    return obj;
+  };
+
+  // Get value at path in some object.
+  util.getIn = function (obj, path) {
+    if (_.isString(path)) {
+      path = [path];
+    }
+    if (path.length === 0) {
+      return obj;
+    }
+    if (_.isObject(obj) && path[0] in obj) {
+      return util.getIn(obj[path[0]], path.slice(1));
+    }
+    return null;
+  };
+
+  // Append to array at path in some object.
+  util.appendIn = function (obj, path, value) {
+    var subObj = util.getIn(obj, path);
+    if (_.isArray(subObj)) {
+      subObj.push(value);
+    }
+    return obj;
+  };
+
+  // Swap two keys at path in some object.
+  util.moveIn = function (obj, path, fromKey, toKey) {
+    var subObj = util.getIn(obj, path);
+    if (_.isArray(subObj)) {
+      if (_.isNumber(fromKey) && _.isNumber(toKey)) {
+        var fromIndex = fromKey;
+        var toIndex = toKey;
+        if (fromIndex !== toIndex &&
+          fromIndex >= 0 && fromIndex < subObj.length &&
+          toIndex >= 0 && toIndex < subObj.length
+        ) {
+          subObj.splice(toIndex, 0, subObj.splice(fromIndex, 1)[0]);
+        }
+      }
+    } else {
+      subObj[toKey] = subObj[fromKey];
+      delete subObj[fromKey];
+    }
+    return obj;
+  };
+
+  // Copy obj, leaving non-JSON behind.
+  util.copyValue = function (value) {
+    return JSON.parse(JSON.stringify(value));
+  };
+
+  // Copy obj recursing deeply.
+  util.deepCopy = function (obj) {
+    if (_.isArray(obj)) {
+      return obj.map(function (item) {
+        return util.deepCopy(item);
+      });
+    } else if (_.isObject(obj)) {
+      var copy = {};
+      _.each(obj, function (value, key) {
+        copy[key] = util.deepCopy(value);
+      });
+      return copy;
+    } else {
+      return obj;
+    }
+  };
+
+  // Check if item matches some value, based on the item's `match` property.
+  util.itemMatchesValue = function (item, value) {
+    var match = item.match;
+    if (!match) {
+      return true;
+    }
+    return _.every(_.keys(match), function (key) {
+      return _.isEqual(match[key], value[key]);
+    });
+  };
+
+  // Create a field definition from a value.
+  util.fieldDefFromValue = function (value) {
+    var def = {
+      type: 'json'
+    };
+    if (_.isString(value)) {
+      def = {
+        type: 'string'
+      };
+    } else if (_.isNumber(value)) {
+      def = {
+        type: 'number'
+      };
+    } else if (_.isBoolean(value)) {
+      def = {
+        type: 'boolean'
+      };
+    } else if (_.isArray(value)) {
+      var arrayItemFields = value.map(function (value, i) {
+        var childDef = util.fieldDefFromValue(value);
+        childDef.key = i;
+        return childDef;
+      });
+      def = {
+        type: 'array',
+        fields: arrayItemFields
+      };
+    } else if (_.isObject(value)) {
+      var objectItemFields = Object.keys(value).map(function (key) {
+        var childDef = util.fieldDefFromValue(value[key]);
+        childDef.key = key;
+        childDef.label = util.humanize(key);
+        return childDef;
+      });
+      def = {
+        type: 'object',
+        fields: objectItemFields
+      };
+    } else if (_.isNull(value)) {
+      def = {
+        type: 'null'
+      };
+    }
+    return def;
+  };
+
+  if (plugin.config.humanize) {
+    // Get the humanize function from a plugin if provided.
+    util.humanize = plugin.require(plugin.config.humanize);
+  } else {
+    // Convert property keys to "human" labels. For example, 'foo' becomes
+    // 'Foo'.
+    util.humanize = function(property) {
+      property = property.replace(/\{\{/g, '');
+      property = property.replace(/\}\}/g, '');
+      return property.replace(/_/g, ' ')
+        .replace(/(\w+)/g, function(match) {
+          return match.charAt(0).toUpperCase() + match.slice(1);
+        });
+    };
+  }
+
+  // Join multiple CSS class names together, ignoring any that aren't there.
+  util.className = function () {
+
+    var classNames = Array.prototype.slice.call(arguments, 0);
+
+    classNames = classNames.filter(function (name) {
+      return name;
+    });
+
+    return classNames.join(' ');
+  };
+
+  // Join keys together to make single "meta" key. For looking up metadata in
+  // the metadata part of the store.
+  util.joinMetaKeys = function (keys) {
+    return keys.join('::');
+  };
+
+  // Split a joined key into separate key parts.
+  util.splitMetaKey = function (key) {
+    return key.split('::');
+  };
+
+  util.metaCacheKey = function (source, params) {
+    params = params || {};
+    return source + '::params(' + JSON.stringify(params) + ')';
+  };
+
+  util.metaErrorCacheKey = function (source, params) {
+    params = params || {};
+    return source + '::params(' + JSON.stringify(params) + ')::error';
+  };
+
+  // Wrap a text value so it has a type. For parsing text with tags.
+  var textPart = function (value, type) {
+    type = type || 'text';
+    return {
+      type: type,
+      value: value
+    };
+  };
+
+  // Parse text that has tags like {{tag}} into text and tags.
+  util.parseTextWithTags = function (value) {
+    value = value || '';
+    var parts = value.split(/{{(?!{)/);
+    var frontPart = [];
+    if (parts[0] !== '') {
+      frontPart = [
+        textPart(parts[0])
+      ];
+    }
+    parts = frontPart.concat(
+      parts.slice(1).map(function (part) {
+        if (part.indexOf('}}') >= 0) {
+          return [
+            textPart(part.substring(0, part.indexOf('}}')), 'tag'),
+            textPart(part.substring(part.indexOf('}}') + 2))
+          ];
+        } else {
+          return textPart('{{' + part, 'text');
+        }
+      })
+    );
+    return [].concat.apply([], parts);
+  };
+
+  // Copy all computed styles from one DOM element to another.
+  util.copyElementStyle = function (fromElement, toElement) {
+    var fromStyle = window.getComputedStyle(fromElement, '');
+
+    if (fromStyle.cssText !== '') {
+      toElement.style.cssText = fromStyle.cssText;
+      return;
+    }
+
+    var cssRules = [];
+    for (var i = 0; i < fromStyle.length; i++) {
+      //console.log(i, fromStyle[i], fromStyle.getPropertyValue(fromStyle[i]))
+      //toElement.style[fromStyle[i]] = fromStyle.getPropertyValue(fromStyle[i]);
+      cssRules.push(fromStyle[i] + ':' + fromStyle.getPropertyValue(fromStyle[i]) + ';');
+    }
+    var cssText = cssRules.join('');
+
+    toElement.style.cssText = cssText;
+  };
+
+  // Object to hold browser sniffing info.
+  var browser = {
+    isChrome: false,
+    isMozilla: false,
+    isOpera: false,
+    isIe: false,
+    isSafari: false
+  };
+
+  // Sniff the browser.
+  var ua = navigator.userAgent;
+  if(ua.indexOf('Chrome') > -1) {
+    browser.isChrome = true;
+  } else if (ua.indexOf('Safari') > -1) {
+    browser.isSafari = true;
+  } else if (ua.indexOf('Opera') > -1) {
+    browser.isOpera = true;
+  } else if (ua.indexOf('Firefox') > -1) {
+    browser.isMozilla = true;
+  } else if (ua.indexOf('MSIE') > -1) {
+    browser.isIe = true;
+  }
+
+  util.browser = browser;
+
+};
