+++ lib/default-config.js
@@ -1,6 +1,7 @@
 'use strict';
 
 var React = require('react/addons');
+var _ = require('underscore');
 
 var utils = require('./utils');
 
@@ -15,21 +16,25 @@ var createObject = function () {
 module.exports = {
 
   // Field element factories
-  createElement_Fields: React.createFactory(require('./components/fields-field')),
-  createElement_Text: React.createFactory(require('./components/text-field')),
-  createElement_Unicode: React.createFactory(require('./components/unicode-field')),
-  createElement_Unknown: React.createFactory(require('./components/unknown-field')),
-  createElement_Copy: React.createFactory(require('./components/copy-field')),
+  createElement_Fields: React.createFactory(require('./components/fields/fields')),
+  createElement_Text: React.createFactory(require('./components/fields/text')),
+  createElement_Unicode: React.createFactory(require('./components/fields/unicode')),
+  createElement_Select: React.createFactory(require('./components/fields/select')),
+  createElement_PrettyTextarea: React.createFactory(require('./components/fields/pretty-textarea')),
+  createElement_UnknownField: React.createFactory(require('./components/fields/unknown')),
+  createElement_Copy: React.createFactory(require('./components/fields/copy')),
 
   // Other element factories
-  createElement_Field: React.createFactory(require('./components/field')),
-  createElement_Label: React.createFactory(require('./components/label')),
-  createElement_Help: React.createFactory(require('./components/help')),
+  createElement_Field: React.createFactory(require('./components/helpers/field')),
+  createElement_Label: React.createFactory(require('./components/helpers/label')),
+  createElement_Help: React.createFactory(require('./components/helpers/help')),
+  createElement_Choices: React.createFactory(require('./components/helpers/choices')),
 
   // Field default values
   getDefaultValue_Fields: createObject,
   getDefaultValue_Text: createString,
   getDefaultValue_Unicode: createString,
+  getDefaultValue_Select: createString,
 
   hasElementFactory: function (name) {
     var config = this;
@@ -41,7 +46,7 @@ module.exports = {
     var config = this;
 
     if (!props.config) {
-      throw new Error('Must pass config in props for: ' + name);
+      props = _.extend({}, props, {config: config});
     }
 
     name = config.getElementName(name);
@@ -51,7 +56,9 @@ module.exports = {
     }
 
     if (name !== 'Unknown') {
-      return config.createElement('Unknown', props, children);
+      if (config.hasElementFactory('Unknown')) {
+        return config.createElement('Unknown', props, children);
+      }
     }
 
     throw new Error('Factory not found for: ' + name);
@@ -73,20 +80,29 @@ module.exports = {
     return utils.dashToPascal(name);
   },
 
-  aliases: {
-  },
+  alias_Dict: 'Object',
 
   // Field helpers
   getFieldTypeName: function (field) {
     var config = this;
 
-    var fieldType = field.type;
+    var fieldType = utils.dashToPascal(field.type);
 
-    if (config.aliases[fieldType]) {
-      fieldType = config.aliases[fieldType];
+    var alias = config['alias_' + fieldType];
+
+    if (alias) {
+      if (_.isFunction(alias)) {
+        return alias(field);
+      } else {
+        return alias;
+      }
     }
 
-    return utils.dashToPascal(fieldType);
+    if (field.list) {
+      fieldType = 'List';
+    }
+
+    return fieldType;
   },
   getFieldDefaultValue: function (field) {
     var config = this;
@@ -100,12 +116,74 @@ module.exports = {
     return '';
   },
   getFieldChoices: function (field) {
+    var config = this;
 
+    return config.normalizeChoices(field.choices);
+  },
+  getFieldReplaceChoices: function (field) {
+    var config = this;
+
+    return config.normalizeChoices(field.replaceChoices);
   },
   getFieldLabel: function (field) {
     return field.label;
   },
   getFieldHelpText: function (field) {
     return field.help_text_html || field.help_text || field.helpText || field.helpTextHtml;
+  },
+
+  // Other helpers
+  humanize: function(property) {
+    property = property.replace(/\{\{/g, '');
+    property = property.replace(/\}\}/g, '');
+    return property.replace(/_/g, ' ')
+    .replace(/(\w+)/g, function(match) {
+      return match.charAt(0).toUpperCase() + match.slice(1);
+    });
+  },
+
+  normalizeChoices: function (choices) {
+    var config = this;
+
+    if (!choices) {
+      return [];
+    }
+
+    // Convert comma separated string to array of strings.
+    if (_.isString(choices)) {
+      choices = choices.split(',');
+    }
+
+    // Convert object to array of objects with `value` and `label` properties.
+    if (!_.isArray(choices) && _.isObject(choices)) {
+      choices = Object.keys(choices).map(function (key) {
+        return {
+          value: key,
+          label: choices[key]
+        };
+      });
+    }
+
+    // Copy the array of choices so we can manipulate them.
+    choices = choices.slice(0);
+
+    // Array of choice arrays should be flattened.
+    choices = _.flatten(choices);
+
+    choices.forEach(function (choice, i) {
+      // Convert any string choices to objects with `value` and `label`
+      // properties.
+      if (_.isString(choice)) {
+        choices[i] = {
+          value: choice,
+          label: config.humanize(choice)
+        };
+      }
+      if (!choices[i].label) {
+        choices[i].label = config.humanize(choices[i].value);
+      }
+    });
+
+    return choices;
   }
 };
