+++ lib/formatic.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var React = require('react/addons');
-//var R = React.DOM;
+var R = React.DOM;
 var _ = require('underscore');
 
 var utils = require('./utils');
@@ -46,7 +46,9 @@ var FormaticControlledClass = React.createClass({
       throw new Error('You must supply a value to the root Formatic component.');
     }
 
-    return config.createField({config: config, field: field, value: value, onChange: this.onChange});
+    return R.div({className: 'formatic'},
+      config.createField({config: config, field: field, value: value, onChange: this.onChange})
+    );
   }
 
 });
