+++ lib/mixins/resize.js
@@ -89,31 +89,28 @@ var onResize = function (ref, fn) {
   fn(ref);
 };
 
-module.exports = function (plugin) {
+module.exports = {
 
-  plugin.exports = {
-
-    componentDidMount: function () {
-      if (this.onResizeWindow) {
-        window.addEventListener('resize', this.onResizeWindow);
-      }
-      this.resizeElementRefs = {};
-    },
-
-    componentWillUnmount: function () {
-      if (this.onResizeWindow) {
-        window.removeEventListener('resize', this.onResizeWindow);
-      }
-      Object.keys(this.resizeElementRefs).forEach(function (ref) {
-        removeResizeIntervalHandlers(this.refs[ref].getDOMNode());
-      }.bind(this));
-    },
-
-    setOnResize: function (ref, fn) {
-      if (!this.resizeElementRefs[ref]) {
-        this.resizeElementRefs[ref] = true;
-      }
-      addResizeIntervalHandler(this.refs[ref].getDOMNode(), onResize.bind(this, ref, fn));
+  componentDidMount: function () {
+    if (this.onResizeWindow) {
+      window.addEventListener('resize', this.onResizeWindow);
     }
-  };
+    this.resizeElementRefs = {};
+  },
+
+  componentWillUnmount: function () {
+    if (this.onResizeWindow) {
+      window.removeEventListener('resize', this.onResizeWindow);
+    }
+    Object.keys(this.resizeElementRefs).forEach(function (ref) {
+      removeResizeIntervalHandlers(this.refs[ref].getDOMNode());
+    }.bind(this));
+  },
+
+  setOnResize: function (ref, fn) {
+    if (!this.resizeElementRefs[ref]) {
+      this.resizeElementRefs[ref] = true;
+    }
+    addResizeIntervalHandler(this.refs[ref].getDOMNode(), onResize.bind(this, ref, fn));
+  }
 };
