+++ lib/mixins/undo-stack.js
@@ -8,7 +8,7 @@ Gives your component an undo stack.
 
 'use strict';
 
-var UndoStack = {
+module.exports = {
   getInitialState: function() {
     return {undo: [], redo: []};
   },
@@ -61,8 +61,4 @@ var UndoStack = {
     this.setStateSnapshot(snapshot);
     this.setState({undo:undo, redo:redo});
   }
-};
-
-module.exports = function (plugin) {
-  plugin.exports = UndoStack;
 };
