+++ lib/plugins/bootstrap.js
@@ -6,26 +6,24 @@ var _ = require('underscore');
 
 var modifiers = {
 
-  'field': {className: 'form-group'},
-  'help': {className: 'help-block'},
-  'sample': {className: 'help-block'},
-  'text': {className: 'form-control'},
-  'textarea': {className: 'form-control'},
-  'pretty-textarea': {className: 'form-control'},
-  'json': {className: 'form-control'},
-  'select': {className: 'form-control'},
-  //'list': {className: 'well'},
-  'list-control': {className: 'form-inline'},
-  'list-item': {className: 'well'},
-  'item-choices': {className: 'form-control'},
-  'add-item': {className: 'glyphicon glyphicon-plus', label: ''},
-  'remove-item': {className: 'glyphicon glyphicon-remove', label: ''},
-  'move-item-back': {className: 'glyphicon glyphicon-arrow-up', label: ''},
-  'move-item-forward': {className: 'glyphicon glyphicon-arrow-down', label: ''},
-  'object-item-key': {className: 'form-control'},
+  'Field': {classes: {'form-group': true}},
+  'Help': {classes: {'help-block': true}},
+  'Sample': {classes: {'help-block': true}},
+  'ListControl': {classes: {'form-inline': true}},
+  'ListItem': {classes: {'well': true}},
+  'ItemChoices': {classes: {'form-control': true}},
+  'AddItem': {classes: {'glyphicon glyphicon-plus': true}, label: ''},
+  'RemoveItem': {classes: {'glyphicon glyphicon-remove': true}, label: ''},
+  'MoveItemBack': {classes: {'glyphicon glyphicon-arrow-up': true}, label: ''},
+  'MoveItemForward': {classes: {'glyphicon glyphicon-arrow-down': true}, label: ''},
+  'ObjectItemKey': {classes: {'form-control': true}},
 
   'Unicode': {classes: {'form-control': true}},
-  'Text': {classes: {'form-control': true}}
+  'Text': {classes: {'form-control': true}},
+  'PrettyTextarea': {classes: {'form-control': true}},
+  'Json': {classes: {'form-control': true}},
+  'Select': {classes: {'form-control': true}}
+  //'list': {classes: 'well'}
 };
 
 module.exports = function (config) {
@@ -36,10 +34,16 @@ module.exports = function (config) {
 
     name = config.getElementName(name);
 
-    if (name in modifiers) {
+    var modifier = modifiers[name];
+
+    if (modifier) {
       props = _.extend({}, props);
-      props.classes = _.extend({}, props.classes, modifiers[name].classes);
+      props.classes = _.extend({}, props.classes, modifier.classes);
+      if ('label' in modifier) {
+        props.label = modifier.label;
+      }
     }
+
     return defaultCreateElement.call(this, name, props, children);
   };
 };
