+++ lib/utils.js
@@ -36,3 +36,48 @@ utils.dashToPascal = function (s) {
   }
   return dashToPascalCache[s];
 };
+
+// Copy all computed styles from one DOM element to another.
+utils.copyElementStyle = function (fromElement, toElement) {
+  var fromStyle = window.getComputedStyle(fromElement, '');
+
+  if (fromStyle.cssText !== '') {
+    toElement.style.cssText = fromStyle.cssText;
+    return;
+  }
+
+  var cssRules = [];
+  for (var i = 0; i < fromStyle.length; i++) {
+    //console.log(i, fromStyle[i], fromStyle.getPropertyValue(fromStyle[i]))
+    //toElement.style[fromStyle[i]] = fromStyle.getPropertyValue(fromStyle[i]);
+    cssRules.push(fromStyle[i] + ':' + fromStyle.getPropertyValue(fromStyle[i]) + ';');
+  }
+  var cssText = cssRules.join('');
+
+  toElement.style.cssText = cssText;
+};
+
+// Object to hold browser sniffing info.
+var browser = {
+  isChrome: false,
+  isMozilla: false,
+  isOpera: false,
+  isIe: false,
+  isSafari: false
+};
+
+// Sniff the browser.
+var ua = navigator.userAgent;
+if(ua.indexOf('Chrome') > -1) {
+  browser.isChrome = true;
+} else if (ua.indexOf('Safari') > -1) {
+  browser.isSafari = true;
+} else if (ua.indexOf('Opera') > -1) {
+  browser.isOpera = true;
+} else if (ua.indexOf('Firefox') > -1) {
+  browser.isMozilla = true;
+} else if (ua.indexOf('MSIE') > -1) {
+  browser.isIe = true;
+}
+
+utils.browser = browser;
